Bring create and drop arguments to tool input_schema
