+++ mcp_server_snowflake/object_manager/objects.py
@@ -1,7 +1,7 @@
 from typing import Literal
 
 from pydantic import BaseModel, Field
-from snowflake.core import CreateMode, Root
+from snowflake.core import Root
 from snowflake.core.database import Database
 from snowflake.core.schema import Schema
 
@@ -49,30 +49,5 @@ class SnowflakeSchema(ObjectMetadata):
 SnowflakeClasses = [SnowflakeDatabase, SnowflakeSchema]
 
 
-async def create_object(
-    object_type: ObjectMetadata,
-    root: Root,
-    mode: Literal["error_if_exists", "replace", "if_not_exists"] = "error_if_exists",
-):
-    if mode == "error_if_exists":
-        create_mode = CreateMode.error_if_exists
-    elif mode == "replace":
-        create_mode = CreateMode.or_replace
-    elif mode == "if_not_exists":
-        create_mode = CreateMode.if_not_exists
-    else:
-        create_mode = CreateMode.if_not_exists
-    core_object = object_type.get_core_object()
-    core_path = object_type.get_core_path(root=root)
-    return core_path.create(core_object, mode=create_mode)
-
-
-def drop_object(object_type: ObjectMetadata, root: Root, if_exists: bool = False):
-    core_object = object_type.get_core_object()
-    core_path = object_type.get_core_path(root=root)
-    return core_path[core_object.name].drop(if_exists=if_exists)
-
-
 # TODOS:
-# - How to integrate optional params like CreateMode into create and if_exists into drop?
 # - Add additional assertions at the class level if possible
